+++ wp-parsely/src/js/admin-settings.ts
@@ -71,7 +71,7 @@ function selectImage( event: Event ) {
 }
 
 /**
- * Adds the necessary event handlers to the Content Helper tab.
+ * Adds the necessary event handlers to the Content Intelligence tab.
  *
  * @since 3.16.0
  */
@@ -89,9 +89,9 @@ function addContentHelperTabEventHandlers(): void {
 		'input#content_helper_traffic_boost_enabled'
 	);
 
-	// Selector for all fieldsets in the Content Helper section.
+	// Selector for all fieldsets in the Content Intelligence section.
 	const fieldsets = document.querySelectorAll(
-		'div.content-helper-section fieldset'
+		'div.content-intelligence-section fieldset'
 	);
 
 	// Event handlers.
@@ -220,7 +220,7 @@ function addContentHelperTabEventHandlers(): void {
 	function enableAllFormFieldsOnSubmit(): void {
 		document.querySelector( '.wp-admin form[name="parsely"]' )
 			?.addEventListener( 'submit', (): void => {
-				const baseSelector = '.wp-admin .content-helper-section fieldset';
+				const baseSelector = '.wp-admin .content-intelligence-section fieldset';
 
 				document.querySelectorAll( `${ baseSelector }[disabled]` )
 					.forEach( ( fieldset: Element ): void => {
@@ -243,7 +243,7 @@ function addContentHelperTabEventHandlers(): void {
 }
 
 /**
- * Displays any messages needed under the Content Helper section.
+ * Displays any messages needed under the Content Intelligence section.
  *
  * @since 3.19.0
  */
@@ -270,19 +270,19 @@ async function displayContentHelperSectionMessages(): Promise<void> {
 
 		if ( err instanceof ContentHelperError ) {
 			if ( ContentHelperErrorCode.PluginSettingsApiSecretNotSet === err.code ) {
-				message = sprintf( '<p><strong>%s</strong></p>', escapeHTML( __( 'All Content Helper AI functionality is disabled because an API Secret has not been set.', 'wp-parsely' ) ) );
+				message = sprintf( '<p><strong>%s</strong></p>', escapeHTML( __( 'All Content Intelligence AI functionality is disabled because an API Secret has not been set.', 'wp-parsely' ) ) );
 			}
 		}
 	} finally {
 		if ( authResponse ) {
 			if ( 200 !== authResponse.api.code ) {
 				const requestAccessLink = sprintf( '<a href="%1$s" target="_blank" rel="noopener">%2$s</a>', 'https://wpvip.com/content-helper/#content-helper-form', __( 'Request access here', 'wp-parsely' ) );
-				/* translators: %s: Link to request access to Content Helper AI functionality. */
-				const messageWithAccessLink = sprintf( escapeHTML( __( 'All Content Helper AI functionality is disabled for this website. %s.', 'wp-parsely' ) ), requestAccessLink );
+				/* translators: %s: Link to request access to Content Intelligence AI functionality. */
+				const messageWithAccessLink = sprintf( escapeHTML( __( 'All Content Intelligence AI functionality is disabled for this website. %s.', 'wp-parsely' ) ), requestAccessLink );
 				message = sprintf( '<p><strong>%s</strong></p>', messageWithAccessLink );
 			} else if ( 200 === authResponse.api.code && 200 !== authResponse.traffic_boost.code ) {
 				const contactSupportLink = sprintf( '<a href="%1$s">%2$s</a>', 'mailto:support@parsely.com', 'support@parsely.com' );
-				/* translators: %s: Link to request access to Content Helper AI functionality. */
+				/* translators: %s: Link to request access to Content Intelligence AI functionality. */
 				const messageWithAccessLink = sprintf( escapeHTML( __( 'Engagement Boost functionality is disabled for this website. To enable it, contact %s.', 'wp-parsely' ) ), contactSupportLink );
 				message = sprintf( '<p><strong>%s</strong></p>', messageWithAccessLink );
 			}
@@ -292,7 +292,7 @@ async function displayContentHelperSectionMessages(): Promise<void> {
 			const div = document.createElement( 'div' );
 			div.className = 'content-helper-message notice notice-error';
 			div.innerHTML = message;
-			const contentHelperSection = document.querySelector( '.content-helper-section' );
+			const contentHelperSection = document.querySelector( '.content-intelligence-section' );
 			if ( contentHelperSection ) {
 				contentHelperSection.insertBefore( div, contentHelperSection.firstChild );
 			}