+++ wp-parsely/src/rest-api/content-helper/class-endpoint-traffic-boost.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * Endpoint: Traffic Boost
- * Parse.ly Content Helper `/traffic-boost` API endpoint class
+ * Parse.ly Content Intelligence `/traffic-boost` API endpoint class
  *
  * @package Parsely
  * @since   3.19.0
@@ -46,7 +46,7 @@ class Endpoint_Traffic_Boost extends Base_Endpoint {
 	 *
 	 * @since 3.19.0
 	 *
-	 * @param Content_Helper_Controller $controller The Content Helper controller.
+	 * @param Content_Helper_Controller $controller The Content Intelligence controller.
 	 */
 	public function __construct( Content_Helper_Controller $controller ) {
 		parent::__construct( $controller );